Move HandlerType to spin-http
