+++ crates/http/Cargo.toml
@@ -16,6 +16,7 @@ serde = { workspace = true }
 spin-app = { path = "../app", optional = true }
 spin-locked-app = { path = "../locked-app" }
 tracing = { workspace = true }
+wasmtime = { workspace = true }
 wasmtime-wasi-http = { workspace = true, optional = true }
 
 [dev-dependencies]
